+++ tests/Feature/Api/ImageUpload/AnonymousTest.php
@@ -1,0 +1,64 @@
+<?php
+
+namespace Tests\Feature\Api\ImageUpload;
+
+use App\Models\ImageUpload;
+use App\Models\User;
+use Illuminate\Foundation\Testing\RefreshDatabase;
+use Illuminate\Foundation\Testing\WithFaker;
+use Illuminate\Http\UploadedFile;
+use Illuminate\Support\Facades\Event;
+use Illuminate\Support\Facades\Storage;
+use Symfony\Component\Routing\Exception\RouteNotFoundException;
+use Tests\TestCase;
+
+class AnonymousTest extends TestCase
+{
+    use RefreshDatabase, WithFaker;
+
+    protected ?User $user = null;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+        Storage::fake('local');
+        Event::fake();
+    }
+
+    public function test_api_authentication_index_forbids_anonymous_access(): void
+    {
+        $response = $this->getJson(route('image-upload.index'));
+        $response->assertUnauthorized();
+    }
+
+    public function test_api_authentication_show_forbids_anonymous_access(): void
+    {
+        $imageUpload = ImageUpload::factory()->create();
+        $response = $this->getJson(route('image-upload.show', $imageUpload->id));
+        $response->assertUnauthorized();
+    }
+
+    public function test_api_authentication_store_forbids_anonymous_access(): void
+    {
+        $response = $this->postJson(route('image-upload.store'), [
+            'file' => UploadedFile::fake()->image('test.jpg'),
+        ]);
+        $response->assertUnauthorized();
+    }
+
+    public function test_api_route_update_is_not_found(): void
+    {
+        $this->expectException(RouteNotFoundException::class);
+
+        $this->putJson(route('image-upload.update', 'non-existent-id'), [
+            'file' => UploadedFile::fake()->image('updated.jpg'),
+        ]);
+    }
+
+    public function test_api_authentication_destroy_forbids_anonymous_access(): void
+    {
+        $imageUpload = ImageUpload::factory()->create();
+        $response = $this->deleteJson(route('image-upload.destroy', $imageUpload->id));
+        $response->assertUnauthorized();
+    }
+}
